Move package information from cacoo-pkg.el to cacoo.el
"<name>-pkg.el" files are not supposed to be checked into packages' Git
repositories.  Package maintainers should provide package information
exclusively in the main library, "<name>.el".

It is the responsibility of package archives (such as MELPA) to extract
that information from there, and to put it in a _newly generated_
"<name>-pkg.el".

Tracking "<name>-pkg.el" needlessly adds ambiguity because it adds a
second, potentially conflicting, sources of truth.